Refactoring Wave 1
