+++ src/gui/device_controls/devicecolorwidget.cpp
