+++ src/gui/device_controls/devicecolorwidget.h
@@ -2,7 +2,7 @@
 #define DEVICECOLORWIDGET_H
 
 #include <QWidget>
-#include "mainwindow.h"
+#include "../mainwindow.h"
 
 namespace Ui {
 class DeviceColorWidget;
