+++ src/gui/device_controls/devicecolorwidget.ui
