+++ src/gui/device_controls/deviceinfowidget.cpp
