+++ src/gui/device_controls/deviceinfowidget.ui
