+++ src/gui/device_controls/devicepowerwidget.cpp
