+++ src/gui/device_options/deviceoptionswidget.cpp
