+++ src/gui/device_options/deviceoptionswidget.h
