+++ src/gui/device_options/fullinfowidget.cpp
