+++ src/gui/device_options/fullinfowidget.h
