+++ src/gui/device_options/fullinfowidget.ui
