+++ src/gui/device_options/mqttconfigwidget.ui
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <ui version="4.0">
- <class>DeviceMQTTServersWidget</class>
- <widget class="QWidget" name="DeviceMQTTServersWidget">
+ <class>MQTTConfigWidget</class>
+ <widget class="QWidget" name="MQTTConfigWidget">
   <property name="geometry">
    <rect>
     <x>0</x>
