+++ src/gui/device_options/setoptionwidget.cpp
