+++ src/gui/device_options/setoptionwidget.h
