+++ src/gui/device_options/setoptionwidget.ui
