+++ src/gui/device_options/wifinetworkswidget.ui
