+++ src/gui/mainwindow.h
@@ -12,7 +12,7 @@
 #include "preferencesdialog.h"
 #include "terminaldialog.h"
 #include "selectserverdialog.h"
-#include "deviceoptionswidget.h"
+#include "device_options/deviceoptionswidget.h"
 
 #include "app/devicemanager.h"
 #include "app/mqttservermanager.h"
